# src/main/resources/api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Letras y Papeles API",
    "description": "Endpoints para la implementación del proyecto Letras y Papeles",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/product/{id}": {
      "get": {
        "tags": [
          "** Endpoints Producto **"
        ],
        "summary": "Obtiene un producto por su ID",
        "description": "Este endpoint devuelve un producto específico basado en su ID",
        "operationId": "getProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Producto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "** Endpoints Producto **"
        ],
        "summary": "Actualiza un producto existente",
        "description": "Este endpoint permite actualizar un producto existente en el sistema",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del producto a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Datos del producto a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Producto actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "** Endpoints Producto **"
        ],
        "summary": "Elimina un producto por su ID",
        "description": "Este endpoint permite eliminar un producto específico basado en su ID",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del producto",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Producto eliminado exitosamente",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Producto no encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/product": {
      "get": {
        "tags": [
          "** Endpoints Producto **"
        ],
        "summary": "Obtiene una lista de todlos los productos",
        "description": "Este endpoint devuelve una lista completa de todas los productos registradas en el sistema",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Lista de productos obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          },
          "204": {
            "description": "No hay productos disponibles (La lista está vacía)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelProductResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "** Endpoints Producto **"
        ],
        "summary": "Crea un nuevo producto",
        "description": "Este endpoint permite crear un nuevo producto en el sistema",
        "operationId": "createProduct",
        "requestBody": {
          "description": "Datos del producto a crear",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              },
              "examples": {
                "Producto ejemplo": {
                  "summary": "Ejemplo de un producto completo",
                  "description": "Producto ejemplo",
                  "value": {
                    "name": "Producto Ejemplo",
                    "sku": "SKU12345",
                    "details": "Detalles del producto",
                    "price": 100
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Producto creado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProductRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "details": {
            "type": "string"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "details": {
            "type": "string"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "EntityModelProductResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sku": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "details": {
            "type": "string"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "CollectionModelEntityModelProductResponse": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "productResponseList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelProductResponse"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    }
  }
}